+++ include/ck_tile/ops/gemm_quant/pipeline/tile_gemm_quant_traits.hpp
@@ -12,8 +12,21 @@ enum struct QuantType : std::uint16_t
 {
     AQuantGrouped = 0,
     BQuantGrouped = 1,
-    RowColQuant   = 2
+    RowColQuant   = 2,
+    TensorQuant   = 3
 };
+
+std::string quant_type_to_string(QuantType quant_type)
+{
+    switch(quant_type)
+    {
+    case QuantType::AQuantGrouped: return "AQuantGrouped";
+    case QuantType::BQuantGrouped: return "BQuantGrouped";
+    case QuantType::RowColQuant: return "RowColQuant";
+    case QuantType::TensorQuant: return "TensorQuant";
+    default: return "Unknown";
+    }
+}
 
 template <bool kPadM_,
           bool kPadN_,